Add "Contributions By Organization"
Adds a contribution by organization page which handily lists all
the contributions from a single organization, auto-populated from the
`data/organizations` directory and based on the
creator/organization/lab/institute fields of every `model` and `simulation.`

Each Org listed this way has the ability to grab the URL to that master
list on the page itself.

Page can be found by the top level navigation under the
"Contribution -> Contributions by Organization" header.

Also made a small change to the entry anchors so URL's can be a bit shorter.
Old URL's will continue to work.

Added the ability have alternate names and spellings for organizations
which are automatically compared against for who contributed what data.
This is all case and whitespace insensitive.