Expanded Simulation Representation and Citations
Expanded the capabilities of the `simulations` entries to support more
general data entries with more than 1 URL to the trajectory. This is
done through the new `trajectory_data` keyword which is exclusive
with `trajectory` + `size`. It allows arbitrary entry of headers and
items, rendering them in smaller tables as part of the larger entry.

The Validation script has been updated to correctly capture this schema.

The simulation page has been updated to show this type of data
while still maintaining the old style without breaking.

New Functionality: collapsed citations in simulations.

Resolved a long outstanding item to make the references and citations
which come as part of the metadata. These start as collapsed items which
can be expanded for space.