Add the 15.14 us Trajectories from Luc-Henri Jolly Jussieu and Jean-Philip Piquemal
Data drawn from its [original site](http://tinker-hp.ip2ct.upmc.fr/?Access-to-the-SARS_Cov2-Protease).
The engineering to make this render correctly is included as part of the
PR this is attached to.

Due to the file sizes, compressing to a single file was not feasible. It
would have been > 1TB.