Prevent entries with only subdomain proteins
Since subdomains are not handled well yet, anyone providing a valid
subdomain, but not the parent protein, to a list of `proteins` will
not have the entry displayed correctly.

This PR adds a check to the validator to prevent subdomain-only entries
while also fixing a couple entries which were subdomain-only so they
will now correctly appear on the site.