Add crossreference with BioExcel Trajectory tools
This PR adds the ability to specify BioExcel Accessions which have
trajectory analysis and modeling through their site at https://bioexcel-cv19.bsc.es/#/

It is a WIP on initial commit as there will need to be some back and forth
to design a pretty interface.

This respects the request and only loads the `iframe` on request, and not
on page load.

Outstanding items:
- Finish adding entries
- Clean up presentation of the entries which are many links
- Resolve conflicts with repeated entries and unique element ID's in page
- Include better titles (maybe by API call) for the entries.
- Optional: Link to the search for the multi-entry items
- Optional: Static image in place before trajectory load